Display different image formats and relative links
Practiced using relative links to link to
the image and extras pages. Displayed mochi in
jpg, gif, svg, and png format using the img tag.
Added the src attribute for the image path and alt
for the alternative text.

This exercise reinforced the best use cases for
each format, how they scale in respect to
dimensions, and the importance of alt text for
accessibility for screen readers.