Explore absolute and relative links in detail
Learned about absolute links and how they connect
to external websites. Learned how relative links
are used to link to other pages on the same
website. Used attributes such as target on
absolute links to manipulate how new pages are
opened. The rel attribute was used with noopener
and noreferrer for security and privacy for pages
that open in a new window.

This exercise was fundamental to understanding
when to use absolute vs relative links and how
their attributes affect behavior and safety.