Create an absolute link and relative link
Linked my Homepage to an external site on a using
an absolute link with href, target, and rel
attributes. Used target="_blank" to open the site
in a new tab, and applied
rel="noopener noreferrer" to improve security and
privacy.

This exercise reinforced the proper security
practices when opening an external page in a new
window.